+++ tests/cli/test_helpers.py
@@ -3,17 +3,19 @@ from textwrap import dedent
 
 import pytest
 
-from ember_mug.cli.helpers import build_sub_rows, print_changes, validate_mac
+from ember_mug import EmberMug
+from ember_mug.cli.helpers import build_sub_rows, print_changes, print_info, print_table, validate_mac
+from ember_mug.consts import LIQUID_STATE_EMPTY, LIQUID_STATE_HEATING
 from ember_mug.data import Change
 
 
-def test_validate_mac():
+def test_validate_mac() -> None:
     with pytest.raises(ArgumentTypeError):
         validate_mac('potato')
     assert validate_mac('9C:DA:8C:19:27:DA') == '9c:da:8c:19:27:da'
 
 
-def test_build_sub_rows():
+def test_build_sub_rows() -> None:
     sub_rows = build_sub_rows(('Test', 'test1, test2, test3', 'test4'))
     assert sub_rows[0][0] == 'Test'
     assert sub_rows[0][1] == 'test1'
@@ -24,18 +26,57 @@ def test_build_sub_rows():
     assert sub_rows[2][1] == 'test3'
 
 
-def test_print_changes(capsys):
-    print_changes(
-        [
-            Change('current_temp', 55.1, 25),
-            Change('target_temp', 50, 25),
-        ],
-        True,
-    )
+def test_print_changes(capsys) -> None:
+    changes = [
+        Change('name', 'Mug Name', 'Test Mug'),
+        Change('liquid_level', 1, 2),
+        Change('liquid_state', LIQUID_STATE_EMPTY, LIQUID_STATE_HEATING),
+        Change('target_temp', 45, 55),
+    ]
+    print_changes(changes, True)
     captured = capsys.readouterr()
     assert captured.out == dedent(
         """\
-        Current Temp changed from 55.10°C to 25.00°C
-        Target Temp changed from 50.00°C to 25.00°C
+        Name changed from "Mug Name" to "Test Mug"
+        Liquid Level changed from "3.33%" to "6.67%"
+        Liquid State changed from "Empty" to "Heating"
+        Target Temp changed from "45.00°C" to "55.00°C"
+        """
+    )
+
+
+def test_print_table(ember_mug: EmberMug, capsys) -> None:
+    assert print_table([]) is None
+    captured = capsys.readouterr()
+    assert captured.out == ''
+
+
+def test_print_info(ember_mug: EmberMug, capsys) -> None:
+    print_info(ember_mug)
+    captured = capsys.readouterr()
+    assert captured.out == dedent(
+        """\
+        Mug Data
+        +--------------+---------+
+        | Mug Name     |         |
+        +--------------+---------+
+        | Meta         | None    |
+        +--------------+---------+
+        | Battery      | None    |
+        +--------------+---------+
+        | Firmware     | None    |
+        +--------------+---------+
+        | LED Colour   | #ffffff |
+        +--------------+---------+
+        | Liquid State | Unknown |
+        +--------------+---------+
+        | Liquid Level | 0.00%   |
+        +--------------+---------+
+        | Current Temp | 0.00°C  |
+        +--------------+---------+
+        | Target Temp  | 0.00°C  |
+        +--------------+---------+
+        | Use Metric   | True    |
+        +--------------+---------+
         """
     )